requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
certifi==2024.7.4
cffi==1.16.0
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
click==8.1.7
constantly==23.10.4
coreapi==2.3.3
coreschema==0.0.4
coverage==7.2.7
cryptography==42.0.8
daphne==4.1.2
Django==4.2.3
django-channels==0.7.0
djangorestframework==3.14.0
djangorestframework-authentication==0.1.1
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.7
h11==0.14.0
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
msgpack==1.0.8
msgpack-python==0.5.6
oauthlib==3.2.2
openapi-codec==1.3.2
packaging==24.1
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.8.0
pyOpenSSL==24.1.0
pytz==2024.1
PyYAML==6.0.1
redis==5.0.7
requests==2.31.0
requests-oauthlib==2.0.0
service-identity==24.1.0
simplejson==3.19.2
six==1.16.0
sqlparse==0.5.0
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
zope.interface==6.4.post2
